+++ scenes/world_locations/garage/garage_scene.gd
@@ -2,7 +2,6 @@ extends WorldLocation
 
 @onready var car_lifts : Node3D = $CarLifts
 @onready var job_car_spots : Node3D = $JobCarSpots
-@onready var pending_cars : Array[JobCar] = []
 @onready var garage_tiles : Node3D = $Tiles
 @onready var player_cars : Node3D = $PlayerCars
 @onready var furniture : Node3D = $Furniture
@@ -36,23 +35,12 @@ func get_job_car_spots() -> Node3D:
 	return job_car_spots
 
 func add_pending_car(pending_car : JobCar):
-	#for car_spot in job_car_spots.get_children():
-		#if car_spot.get_child_count() == 0:
-			#pending_car.connect("start_repair", start_car_repair)
-			#pending_car.connect("repair_completed", finish_car_repair)
-			#pending_car.connect("ask_for_repair", process_pending_car_repair_request)
-			#break
-
-	# TODO: Fix above
-	#print_debug(str(pending_cars.size()) + "/" + str(job_car_spots.get_child_count()))
-	if pending_cars.size() < job_car_spots.get_child_count():
-		pending_cars.append(pending_car)
-		pending_car.connect("start_repair", start_car_repair)
-		pending_car.connect("repair_completed", finish_car_repair)
-		pending_car.connect("ask_for_repair", process_pending_car_repair_request)
-
-func erase_pending_car(car : JobCar) -> void:
-	pending_cars.erase(car)
+	for car_spot in job_car_spots.get_children():
+		if car_spot.get_child_count() == 0:
+			pending_car.connect("start_repair", start_car_repair)
+			pending_car.connect("repair_completed", finish_car_repair)
+			pending_car.connect("ask_for_repair", process_pending_car_repair_request)
+			break
 
 func start_car_repair(car : JobCar, is_started_by_player : bool):
 	var vacant_car_lift : CarLift = find_available_car_lift()
@@ -60,16 +48,15 @@ func start_car_repair(car : JobCar, is_started_by_player : bool):
 	car.global_position = vacant_car_lift.global_position
 	car.global_rotation = vacant_car_lift.global_rotation
 	vacant_car_lift.start(car, is_started_by_player)
-	pending_cars.erase(car)
 
 func finish_car_repair(cash_reward : int, rep_reward : int, is_repaired_by_player : bool):
 	repair_completed.emit(cash_reward, rep_reward, is_repaired_by_player)
 
 func get_pending_cars() -> Array[JobCar]:
-	#var pending_cars : Array[JobCar] = []
-	#for car_spot in job_car_spots.get_children():
-		#if car_spot.get_child_count() != 0:
-			#pending_cars.append(car_spot.get_child(0))
+	var pending_cars : Array[JobCar] = []
+	for car_spot in job_car_spots.get_children():
+		if car_spot.get_child_count() != 0:
+			pending_cars.append(car_spot.get_child(0))
 	return pending_cars
 
 func show_unlockable_tiles():
@@ -143,7 +130,7 @@ func _on_visibility_changed():
 	camera.global_rotation = camera_initial_rotation
 	raycast_default_state.raycast_enabled = self.visible
 
-func remove_car(id : int):
+func remove_car(id : String):
 	for car in player_cars.get_children():
 		if car.internal_id == id:
 			car.queue_free()